review error 수정

+++ src/main/java/com/tt/talktok/entity/Review.java
@@ -2,6 +2,7 @@ package com.tt.talktok.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
+import org.hibernate.annotations.ColumnDefault;
 import org.hibernate.annotations.CreationTimestamp;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
@@ -21,7 +22,7 @@ public class Review {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "rev_no")
-    private Long revNo;
+    private int revNo;
 
     @Column(name = "rev_name")
     private String revName;
